Change to VM to accomodate sections
