+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/moremenu/MoreMenuItemSection.kt
@@ -1,0 +1,30 @@
+package com.woocommerce.android.ui.moremenu
+
+import androidx.annotation.ColorRes
+import androidx.annotation.DimenRes
+import androidx.annotation.DrawableRes
+import androidx.annotation.StringRes
+
+data class MoreMenuItemSection(
+    @StringRes val title: Int?,
+    val items: List<MoreMenuItemButton>,
+)
+
+data class MoreMenuItemButton(
+    @StringRes val title: Int,
+    @StringRes val description: Int,
+    @DrawableRes val icon: Int,
+    val isEnabled: Boolean = true,
+    val badgeState: BadgeState? = null,
+    val onClick: () -> Unit = {},
+)
+
+data class BadgeState(
+    @DimenRes val badgeSize: Int,
+    @ColorRes val backgroundColor: Int,
+    @ColorRes val textColor: Int,
+    val textState: TextState,
+    val animateAppearance: Boolean = false,
+)
+
+data class TextState(val text: String, @DimenRes val fontSize: Int)
